+++ packages/url/package.json
@@ -1,8 +1,9 @@
 {
   "name": "@solid-primitives/url",
   "version": "0.0.100",
-  "description": "A template primitive example.",
-  "author": "Your Name <you@youremail.com>",
+  "description": "Reactive primitives covering Browser's location, URL and URLSearchParams interfaces.",
+  "author": "Damian Tarnawski @thetarnav <gthetarnav@gmail.com>",
+  "contributors": [],
   "license": "MIT",
   "homepage": "https://github.com/davedbase/solid-primitives/tree/main/packages/url#readme",
   "repository": {
@@ -15,7 +16,13 @@
   "primitive": {
     "name": "url",
     "stage": 2,
-    "list": [],
+    "list": [
+      "createLocationState",
+      "createLocationSearchParams",
+      "createSearchParams",
+      "createURL",
+      "createURLRecord"
+    ],
     "category": "Browser APIs"
   },
   "private": false,
@@ -28,12 +35,16 @@
     "dist"
   ],
   "scripts": {
-    "start": "vite serve dev --host",
+    "start": "vite serve --host",
     "dev": "yarn start",
     "build": "tsup",
-    "test": "uvu -r solid-register"
+    "test": "vitest run test",
+    "test:watch": "vitest watch test"
   },
   "keywords": [
+    "URL",
+    "location",
+    "URLSearchParams",
     "solid",
     "primitives"
   ],
@@ -47,14 +58,13 @@
   "devDependencies": {
     "jsdom": "18.1.1",
     "prettier": "^2.5.1",
-    "solid-register": "0.1.5",
     "tslib": "^2.3.1",
-    "typescript": "^4.5.5",
-    "tsup": "^5.11.13",
-    "uvu": "^0.5.3",
-    "unocss": "0.26.2",
-    "vite": "2.8.4",
-    "vite-plugin-solid": "2.2.5"
+    "typescript": "^4.6.2",
+    "tsup": "^5.12.1",
+    "unocss": "0.29.5",
+    "vite": "2.8.6",
+    "vite-plugin-solid": "2.2.6",
+    "vitest": "^0.7.7"
   },
   "peerDependencies": {
     "solid-js": "^1.3.0"
