--- packages/url/test/index.test.ts
@@ -1,3 +1,0 @@
-import { createRoot } from "solid-js";
-import { suite } from "uvu";
-import * as assert from "uvu/assert";
